+++ demo/src/main/java/com/carlos/cutils/demo/AnimActivity.kt
@@ -1,9 +1,13 @@
 package com.carlos.cutils.demo
 
 import android.os.Bundle
+import android.view.View
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import com.carlos.cutils.aop.CSingleClick
+import com.carlos.cutils.extend.lastClickTime
 import com.carlos.cutils.extend.singleClick
+import com.carlos.cutils.util.LogUtils
 
 /**
  * Created by Carlos on 2019-05-27.
@@ -14,15 +18,23 @@ class AnimActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.test11)
 
-        LogUtil.isShowLog = true
+        LogUtils.isShowLog = true
         val textView = findViewById<TextView>(R.id.textview)
         textView.singleClick(3000) { }
+        textView.lastClickTime
+//        textview.setOnClickListener(object : View.OnClickListener {
+//            @CSingleClick
+//            override fun onClick(v: View?) {
+////                TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//            }
+//
+//        })
     }
 
-//    @SingleClick
-//    fun back(view: View) {
+    @CSingleClick
+    fun back(view: View) {
 //        finish()
-//        overridePendingTransition(R.anim.c_left_in, R.anim.c_left_in)
-//    }
+        overridePendingTransition(R.anim.c_left_in, R.anim.c_left_in)
+    }
 
 }